# swagger-spec.yaml
openapi: 3.0.0
paths:
...
  /users:
    post:
      operationId: UserController_create
      summary: Create a new user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserDto"
      responses:
        "201":
          description: User successfully created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateUserDto"
        "400":
          description: Bad Request
      tags: &a3
        - users
      security: &a4
        - bearer: []
    get:
      operationId: UserController_findAll
      summary: Get all users
      parameters: []
      responses:
        "200":
          description: Users found.
      tags: *a3
      security: *a4
  /users/{id}:
    get:
      operationId: UserController_findOne
      summary: Get a user by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        "200":
          description: User found.
        "404":
          description: User not found.
      tags: *a3
      security: *a4
    put:
      operationId: UserController_update
      summary: Update a user by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserDto"
      responses:
        "200":
          description: User successfully updated.
        "404":
          description: User not found.
      tags: *a3
      security: *a4
    delete:
      operationId: UserController_deleteUser
      summary: Delete a user
      parameters:
        - name: id
          required: true
          in: path
          description: The id of the user to delete
          schema:
            type: number
      responses:
        "200":
          description: User deleted.
        "404":
          description: User not found.
      tags: *a3
      security: *a4
  /products:
    post:
      operationId: ProductController_create
      summary: Create a new product
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: The product has been successfully created.
        "403":
          description: Forbidden.
      tags: &a5
        - products
      security: &a6
        - bearer: []
    get:
      operationId: ProductController_getAll
      summary: Get all products with pagination
      parameters:
        - name: page
          required: false
          in: query
          description: Page number for pagination
          schema:
            type: number
        - name: limit
          required: false
          in: query
          description: Number of items per page for pagination
          schema:
            type: number
      responses:
        "200":
          description: Return all products.
      tags: *a5
      security: *a6
  /products/{id}:
    get:
      operationId: ProductController_getOne
      summary: Get a single product by ID
      parameters:
        - name: id
          required: true
          in: path
          description: Product ID
          schema:
            type: number
      responses:
        "200":
          description: Return a single product.
        "404":
          description: Product not found.
      tags: *a5
      security: *a6
    put:
      operationId: ProductController_update
      summary: Update product by ID
      parameters:
        - name: id
          required: true
          in: path
          description: Product ID
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: The product has been successfully updated.
        "403":
          description: Forbidden.
        "404":
          description: Product not found.
      tags: *a5
      security: *a6
    delete:
      operationId: ProductController_delete
      summary: Delete product by ID
      parameters:
        - name: id
          required: true
          in: path
          description: Product ID
          schema:
            type: number
      responses:
        "200":
          description: The product has been successfully deleted.
        "403":
          description: Forbidden.
        "404":
          description: Product not found.
      tags: *a5
      security: *a6
...
info:
  title: Salekhard
  description: Welcome to Salekhard
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    basic:
      type: http
      scheme: basic
  schemas:
...
    CreateUserDto:
      type: object
      properties:
        firstName:
          type: string
          example: Sergey
          description: The first name of the user
        lastName:
          type: string
          example: Nilogov
          description: The last name of the user
        patronymic:
          type: string
          example: Aleksandrovich
          description: The patronymic name of the user (optional)
        gender:
          type: string
          example: male
          description: The gender of the user
        phone:
          type: string
          example: "+79220763378"
          description: The phone number of the user
        email:
          type: string
          example: sergey@example.com
          description: The email address of the user
        password:
          type: string
          example: Password123
          description: The password of the user
      required:
        - firstName
        - lastName
        - patronymic
        - gender
        - phone
        - email
        - password
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
          example: Awesome Product
          description: The name of the product
        description:
          type: string
          example: Detailed description of the product.
          description: The description of the product
        price:
          type: number
          example: 299
          description: The price of the product
        image:
          type: string
          example: /images/product.jpg
          description: The image path of the product
      required:
        - name
        - description
        - price
        - image
...